refactor: 모달 레이아웃 4종을 modal 폴더 하위로 이동
